# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  JWT:
    type: apiKey
    name: Authorization
    in: header
paths:
  /service:
    x-swagger-router-controller: service
    post:
      security:
        - JWT: []
      consumes:
        - multipart/form-data
      description: Creates new service
      operationId: create
      parameters:
        - in: formData
          name: image
          type: file
          description: The file to upload.
        - in: formData
          name: tag
          type: string
        - in: formData
          name: title
          type: string
        - in: formData
          name: description
          type: string
        - in: formData
          name: phone
          type: string
        - in: formData
          name: email
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Service"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /service/{id}:
    x-swagger-router-controller: service
    get:
      security:
        - JWT: []
      description: Get service by id
      operationId: getById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the service to get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Service"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      security:
        - JWT: []
      description: Delete service by id
      operationId: deleteById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the service to delete
      responses:
        '204':
          description: The resource was deleted successfully.
        '500':
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      security:
        - JWT: []
      consumes:
        - multipart/form-data
      description: Updates service
      operationId: updateById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the service to update
        - in: formData
          name: image
          type: file
          description: The file to upload.
        - in: formData
          name: tag
          type: string
        - in: formData
          name: title
          type: string
        - in: formData
          name: description
          type: string
        - in: formData
          name: phone
          type: string
        - in: formData
          name: email
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Service"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /post:
    x-swagger-router-controller: post
    post:
      security:
        - JWT: []
      consumes:
        - multipart/form-data
      description: Creates new post
      operationId: create
      parameters:
        - in: formData
          name: image
          type: file
          description: The file to upload.
        - in: formData
          name: tag
          type: string
        - in: formData
          name: title
          type: string
        - in: formData
          name: description
          type: string
        - in: formData
          name: phone
          type: string
        - in: formData
          name: email
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Post"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /post/{id}:
    x-swagger-router-controller: post
    get:
      security:
        - JWT: []
      description: Get post by id
      operationId: getById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the post to get
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Post"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      security:
        - JWT: []
      description: Delete post by id
      operationId: deleteById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the post to delete
      responses:
        '204':
          description: The resource was deleted successfully.
        '500':
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      security:
        - JWT: []
      consumes:
        - multipart/form-data
      description: Updates post
      operationId: updateById
      parameters:
        - in: path
          name: id
          type: string
          required: true
          description: String ID of the post to update
        - in: formData
          name: image
          type: file
          description: The file to upload.
        - in: formData
          name: tag
          type: string
        - in: formData
          name: title
          type: string
        - in: formData
          name: description
          type: string
        - in: formData
          name: phone
          type: string
        - in: formData
          name: email
          type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Post"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /auth/login:
    x-swagger-router-controller: auth
    post:
      operationId: login
      description: Login
      parameters:
        - in: body
          name: loginParameters
          description: Login credentials
          schema:
            type: object
            required:
              - email
              - password
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /auth/signup:
    x-swagger-router-controller: auth
    post:
      operationId: signup
      description: Signup
      parameters:
        - in: body
          name: signUpParameters
          description: SignUp credentials
          schema:
            type: object
            required:
              - email
              - password
              - phone
            properties:
              email:
                type: string
              password:
                type: string
              phone:
                type: string
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SignUpResponse"
        "500":
          description: Error
          schema:
            $ref: "#/definitions/Error"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /swagger:
    x-swagger-pipe: swagger_raw
definitions:
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  Service:
    properties:
      id:
        type: string
      userId:
        type: string
      tag:
        type: string
      photoUrl:
        type: string
      title:
        type: string
      description:
        type: string
      phone:
        type: string
      email:
        type: string
  Post:
    properties:
      id:
        type: string
      userId:
        type: string
      tag:
        type: string
      photoUrl:
        type: string
      title:
        type: string
      description:
        type: string
      phone:
        type: string
      email:
        type: string
  Error:
    properties:
      data:
        type: object
        properties:
          message:
            type: string
          error:
            type: string
          code:
            type: number
  LoginResponse:
    properties:
      token:
        type: string
  User:
    properties:
      id:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      photoUrl:
        type: string
      createdAt:
        type: string
  SignUpResponse:
    properties:
      user:
        $ref: "#/definitions/User"
      token:
        type: string